backend/recomendations/src/requirements.txt
accelerate==1.10.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
filelock==3.13.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2024.6.1
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
psutil==7.1.0
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
sympy==1.13.3
tokenizers==0.22.1
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3